+++ onediff_sd_webui_extensions/compile/oneflow/mock/sgm.py
@@ -8,6 +8,14 @@ from onediff.infer_compiler.backends.oneflow.transform import proxy_class
 from .common import CrossAttentionOflow, GroupNorm32Oflow, timestep_embedding
 
 
+def cat(tensors, *args, **kwargs):
+    if len(tensors) == 2:
+        a, b = tensors
+        a = flow.nn.functional.interpolate_like(a, like=b, mode="nearest")
+        tensors = (a, b)
+    return flow.cat(tensors, *args, **kwargs)
+
+
 # https://github.com/Stability-AI/generative-models/blob/059d8e9cd9c55aea1ef2ece39abf605efb8b7cc9/sgm/modules/diffusionmodules/openaimodel.py#L816
 class UNetModelOflow(proxy_class(UNetModel)):
     def forward(self, x, timesteps=None, context=None, y=None, **kwargs):
@@ -29,7 +37,8 @@ class UNetModelOflow(proxy_class(UNetModel)):
             hs.append(h)
         h = self.middle_block(h, emb, context)
         for module in self.output_blocks:
-            h = flow.cat([h, hs.pop()], dim=1)
+            # h = flow.cat([h, hs.pop()], dim=1)
+            h = cat([h, hs.pop()], dim=1)
             h = module(h, emb, context)
         h = h.type(x.dtype)
         return self.out(h)
