+++ onediff_sd_webui_extensions/compile/oneflow/mock/vae.py
@@ -1,0 +1,12 @@
+from modules.sd_vae_approx import VAEApprox
+
+from onediff.infer_compiler.backends.oneflow.transform import proxy_class, register
+
+
+class VAEApproxOflow(proxy_class(VAEApprox)):
+    pass
+
+
+torch2oflow_class_map = {
+    VAEApprox: VAEApproxOflow,
+}
