+++ onediff_sd_webui_extensions/compile/oneflow/utils.py
@@ -11,7 +11,8 @@ def init_oneflow_backend():
             "Backend oneflow for OneDiff is invalid, please make sure you have installed OneFlow"
         )
 
-    from .mock import ldm, sgm
+    from .mock import ldm, sgm, vae
 
     register(package_names=["ldm"], torch2oflow_class_map=ldm.torch2oflow_class_map)
     register(package_names=["sgm"], torch2oflow_class_map=sgm.torch2oflow_class_map)
+    register(package_names=["modules"], torch2oflow_class_map=vae.torch2oflow_class_map)
