+++ onediff_sd_webui_extensions/compile/vae.py
@@ -1,29 +1,20 @@
+from compile.utils import get_onediff_backend
 from modules import shared
 from modules.sd_vae_approx import VAEApprox
 from modules.sd_vae_approx import model as get_vae_model
 from modules.sd_vae_approx import sd_vae_approx_models
 
-from onediff.infer_compiler import oneflow_compile
-from onediff.infer_compiler.backends.oneflow.transform import proxy_class, register
+from onediff.infer_compiler import compile
+
+# from compile import get_compiled_graph
 
 __all__ = ["VaeCompileCtx"]
 
 compiled_models = {}
 
 
-class VAEApproxOflow(proxy_class(VAEApprox)):
-    pass
-
-
-torch2oflow_class_map = {
-    VAEApprox: VAEApproxOflow,
-}
-
-register(package_names=["modules"], torch2oflow_class_map=torch2oflow_class_map)
-
-
 class VaeCompileCtx(object):
-    def __init__(self, options=None):
+    def __init__(self, backend=None, options=None):
         self._options = options
         # https://github.com/AUTOMATIC1111/stable-diffusion-webui/blob/75336dfc84cae280036bc52a6805eb10d9ae30ba/modules/sd_vae_approx.py#L43
         self._model_name = (
@@ -32,14 +23,19 @@ class VaeCompileCtx(object):
             else "model.pt"
         )
         self._original_model = get_vae_model()
+        self.backend = backend
 
     def __enter__(self):
+        # TODO: support nexfort here
         if self._original_model is None:
             return
         global compiled_models
         model = compiled_models.get(self._model_name)
+        backend = str(self.backend or get_onediff_backend())
         if model is None:
-            model = oneflow_compile(self._original_model, options=self._options)
+            model = compile(
+                self._original_model, backend=backend, options=self._options
+            )
             compiled_models[self._model_name] = model
         sd_vae_approx_models[self._model_name] = model
